create formSections folder and move sections inside for better readability
